✨ [feat] 기록중 페이지 모달, 기록완료 버튼 분리 (EndRecordingButton)
