+++ apps/web/src/mocks/data/recording.ts
@@ -41,7 +41,7 @@ export const recording: Record<number, RecordingResponse> = {
     matchTime: { hour: 18, minute: 30, second: 0, nano: 0 },
     userId: 2,
     watchCnt: 1,
-    result: 'LOSE',
+    result: 'LOSS',
     baseballTeam: 'SSG_LANDERS',
     positiveEmotionPercent: 40,
     negativeEmotionPercent: 60,
@@ -71,7 +71,7 @@ export const recording: Record<number, RecordingResponse> = {
     matchTime: { hour: 19, minute: 30, second: 0, nano: 0 },
     userId: 3,
     watchCnt: 3,
-    result: 'WIN',
+    result: null,
     baseballTeam: 'LG_TWINS',
     positiveEmotionPercent: 90,
     negativeEmotionPercent: 10,
@@ -101,7 +101,7 @@ export const recording: Record<number, RecordingResponse> = {
     matchTime: { hour: 18, minute: 30, second: 0, nano: 0 },
     userId: 4,
     watchCnt: 0,
-    result: 'DRAW',
+    result: null,
     baseballTeam: 'DOOSAN_BEARS',
     positiveEmotionPercent: 50,
     negativeEmotionPercent: 50,
@@ -131,7 +131,7 @@ export const recording: Record<number, RecordingResponse> = {
     matchTime: { hour: 19, minute: 30, second: 0, nano: 0 },
     userId: 5,
     watchCnt: 5,
-    result: 'WIN',
+    result: null,
     baseballTeam: 'SSG_LANDERS',
     positiveEmotionPercent: 80,
     negativeEmotionPercent: 20,
@@ -161,7 +161,7 @@ export const recording: Record<number, RecordingResponse> = {
     matchTime: { hour: 18, minute: 30, second: 0, nano: 0 },
     userId: 6,
     watchCnt: 4,
-    result: 'LOSE',
+    result: null,
     baseballTeam: 'DOOSAN_BEARS',
     positiveEmotionPercent: 25,
     negativeEmotionPercent: 75,
